refactor(settings): seperate files for ollama and openai
